avoid needless cast

+++ src/tree/node.rs
@@ -331,7 +331,7 @@ impl Node {
 
   /// Get the node type
   pub fn get_type(&self) -> Option<NodeType> {
-    NodeType::from_int(xmlGetNodeType(self.node_ptr()) as u32)
+    NodeType::from_int(xmlGetNodeType(self.node_ptr()))
   }
 
   /// Add a previous sibling
